CAnonizer 3 Release 14th MArch

+++ src/network/api/userApi.ts
@@ -1,6 +1,6 @@
 import { message } from "antd";
 
-import { handleError, isServer } from "../../utils/generalUtility";
+import { getCookies, handleError, isServer } from "../../utils/generalUtility";
 import {
   setAuthToken,
   removeAuthToken,
@@ -10,19 +10,30 @@ import {
   setLogout,
 } from "../../store/slices/authSlice";
 import { showMultiUserModal, updateStatus } from "../../store/slices/uiSlice";
-import { setValue } from "../../store/slices/utilsSlice";
 import NetworkCall from "../networkCall";
 import UserRequest from "../request/userRequest";
 import { store } from "../../store";
-import { setFilterCanonizedTopics } from "../../store/slices/filtersSlice";
+import {
+  setFilterCanonizedTopics,
+  setRemoveFilters,
+} from "../../store/slices/filtersSlice";
 import { setHeaderData } from "src/store/slices/notificationSlice";
 import { setIsChecked } from "src/store/slices/recentActivitiesSlice";
 
 export const createToken = async () => {
   try {
     const token = await NetworkCall.fetch(UserRequest.createToken());
+
+    if (!isServer()) {
+      document.cookie =
+        "loginToken=" +
+        token?.data?.access_token +
+        "; expires=Thu, 15 Jul 2030 00:00:00 UTC; path=/";
+      localStorage.setItem("auth_token", token?.data?.access_token);
+    }
+
     store.dispatch(setAuthToken(token?.data?.access_token));
-    localStorage.setItem("auth_token", token?.data?.access_token);
+
     return token.data;
   } catch (error) {
     handleError(error);
@@ -31,21 +42,21 @@ export const createToken = async () => {
 
 export const login = async (email: string, password: string) => {
   try {
-    const authToken = await createToken();
-
     const res = await NetworkCall.fetch(
-      UserRequest.loginUser(email, password, authToken?.access_token)
+      UserRequest.loginUser(email, password, (getCookies() as any)?.loginToken)
     );
 
-    let payload = {
-      ...res.data.user,
-      token: res.data.auth?.access_token,
-      refresh_token: res.data?.auth?.refresh_token,
-    };
+    store.dispatch(setAuthToken(res.data.auth?.access_token));
+
     document.cookie =
       "loginToken=" +
       res.data.auth?.access_token +
       "; expires=Thu, 15 Jul 2030 00:00:00 UTC; path=/";
+
+    let payload = {
+      ...res.data.user,
+    };
+
     store.dispatch(setLoggedInUser(payload));
 
     return res;
@@ -73,6 +84,21 @@ export const logout = async (error = "", status = null, count: number = 1) => {
       store.dispatch(updateStatus(status));
       store.dispatch(setHeaderData({ count: 0, list: [] }));
       store.dispatch(setIsChecked(false));
+      store.dispatch(
+        setRemoveFilters({
+          page_number: 1,
+          page_size: 15,
+          nameSpace: "/General/",
+          namespace_id: "1",
+          asofdate: Date.now() / 1000,
+          asof: "default",
+          filterByScore: 0,
+          algorithm: "blind_popularity",
+          search: "",
+          includeReview: false,
+          is_archive: 0,
+        })
+      );
 
       if (+state.ui.apiStatus === +status) {
         return;
@@ -84,21 +110,24 @@ export const logout = async (error = "", status = null, count: number = 1) => {
       return true;
     }
 
-    if (!isServer()) {
-      localStorage.setItem("logout_type", "true");
-      store.dispatch(setValue({ label: "logout_type", value: true }));
-    }
-
     let res = await NetworkCall.fetch(UserRequest.logoutCall(auth.token));
 
-    store.dispatch(setLogout());
-    store.dispatch(setIsChecked(false));
-    store.dispatch(logoutUser());
-    store.dispatch(removeAuthToken());
-    document.cookie =
-      "loginToken=; expires=Thu, 15 Jul 2030 00:00:00 UTC; path=/";
-    store.dispatch(setHeaderData({ count: 0, list: [] }));
-    await createToken();
+    if (res?.status_code === 200) {
+      document.cookie =
+        "loginToken=; expires=Thu, 15 Jul 2030 00:00:00 UTC; path=/";
+
+      if (!(getCookies() as any)?.loginToken) {
+        const tRes = await createToken();
+        if (tRes?.access_token) {
+          store.dispatch(setLogout());
+          store.dispatch(setIsChecked(false));
+          store.dispatch(logoutUser());
+          store.dispatch(removeAuthToken());
+          store.dispatch(setHeaderData({ count: 0, list: [] }));
+        }
+      }
+    }
+
     return res;
   } catch (error) {
     store.dispatch(logoutUser());
@@ -110,10 +139,8 @@ export const logout = async (error = "", status = null, count: number = 1) => {
 
 export const register = async (values: object) => {
   try {
-    const authToken = await createToken();
-
     const res = await NetworkCall.fetch(
-      UserRequest.registerUser(values, authToken?.access_token)
+      UserRequest.registerUser(values, (getCookies() as any)?.loginToken)
     );
 
     return res;
@@ -134,17 +161,12 @@ export const register = async (values: object) => {
 
 export const verifyOtp = async (values: object) => {
   try {
-    const authToken = await createToken();
-
-    const res = await NetworkCall.fetch(
-      UserRequest.verifyUser(values, authToken?.access_token)
-    );
+    const res = await NetworkCall.fetch(UserRequest.verifyUser(values));
 
     let payload = {
       ...res.data.user,
-      token: res.data.auth?.access_token,
-      refresh_token: res.data?.auth?.refresh_token,
     };
+
     document.cookie =
       "loginToken=" +
       res.data.auth?.access_token +
@@ -158,11 +180,10 @@ export const verifyOtp = async (values: object) => {
 };
 
 export const changePassword = async (values: object) => {
-  let state = store.getState();
-  const { auth } = state;
+  const cc: any = getCookies();
+
   const res = await NetworkCall.fetch(
-    UserRequest.changePassword(values, auth.loggedInUser.token)
-    //UserRequest.changePassword(values, auth.token)
+    UserRequest.changePassword(values, cc?.loginToken)
   );
   return res;
 };
@@ -186,10 +207,8 @@ export const deleteProfileImage = () => {
 // social login path
 export const socialLogin = async (values: object) => {
   try {
-    const authToken = await createToken();
-
     const res = await NetworkCall.fetch(
-      UserRequest.userSocialLogin(values, authToken?.access_token)
+      UserRequest.userSocialLogin(values, (getCookies() as any)?.loginToken)
     );
 
     return res;
@@ -199,18 +218,9 @@ export const socialLogin = async (values: object) => {
 };
 
 export const socialLoginCallback = async (values: object, router) => {
-  const state = store.getState();
-
   try {
-    let token = null;
-    let authToken = null;
-
-    if (state.auth.token) {
-      token = state.auth.token;
-    } else {
-      authToken = await createToken();
-      token = authToken?.access_token;
-    }
+    const cc: any = getCookies();
+    let token = cc?.loginToken;
 
     const res = await NetworkCall.fetch(
       UserRequest.userSocialLoginCallback(values, token)
@@ -218,8 +228,6 @@ export const socialLoginCallback = async (values: object, router) => {
 
     let payload = {
       ...res.data.user,
-      token: res.data.auth?.access_token,
-      refresh_token: res.data?.auth?.refresh_token,
     };
     document.cookie = "loginToken=" + res.data.auth?.access_token + ";path=/";
     store.dispatch(setLoggedInUser(payload));
@@ -260,9 +268,8 @@ export const socialLoginCallback = async (values: object, router) => {
 
 export const getCountryCodes = async () => {
   try {
-    const authToken = await createToken();
     const res = await NetworkCall.fetch(
-      UserRequest.getCountryCodes(authToken?.access_token)
+      UserRequest.getCountryCodes((getCookies() as any)?.loginToken)
     );
 
     return res;
@@ -278,7 +285,7 @@ export const GetUserProfileInfo = async (token = "") => {
   if (token) {
     tcn = token;
   } else {
-    tcn = auth.loggedInUser?.token;
+    tcn = auth?.token;
   }
   const res = await NetworkCall.fetch(UserRequest.GetUserProfileInfo(tcn))
     .then((value) => {
@@ -294,7 +301,7 @@ export const UpdateUserProfileInfo = async (values: object) => {
   let state = store.getState();
   const { auth } = state;
   const res = await NetworkCall.fetch(
-    UserRequest.UpdateUserProfileInfo(values, auth.loggedInUser.token)
+    UserRequest.UpdateUserProfileInfo(values, auth?.token)
   )
     .then((value) => {
       let payload = {
@@ -304,14 +311,7 @@ export const UpdateUserProfileInfo = async (values: object) => {
         phone_number: value.data.phone_number,
         birthday: value.data.birthday,
         email: value.data.email,
-        token: auth.loggedInUser.token,
-        refresh_token: auth.loggedInUser.refresh_token,
       };
-      document.cookie =
-        "loginToken=" +
-        value?.data.auth?.access_token +
-        "; expires=Thu, 15 Jul 2030 00:00:00 UTC; path=/";
-
       store.dispatch(setLoggedInUser(payload));
       return value;
     })
@@ -324,9 +324,7 @@ export const UpdateUserProfileInfo = async (values: object) => {
 export const GetMobileCarrier = async () => {
   let state = store.getState();
   const { auth } = state;
-  const res = await NetworkCall.fetch(
-    UserRequest.GetMobileCarrier(auth.loggedInUser?.token)
-  )
+  const res = await NetworkCall.fetch(UserRequest.GetMobileCarrier(auth?.token))
     .then((value) => {
       return value;
     })
@@ -339,9 +337,7 @@ export const GetMobileCarrier = async () => {
 export const SendOTP = async (values: object) => {
   let state = store.getState();
   const { auth } = state;
-  const res = await NetworkCall.fetch(
-    UserRequest.SendOTP(values, auth.loggedInUser.token)
-  )
+  const res = await NetworkCall.fetch(UserRequest.SendOTP(values, auth.token))
     .then((value) => {
       return value;
     })
@@ -355,7 +351,7 @@ export const VerifyOTP = async (values: object) => {
   let state = store.getState();
   const { auth } = state;
   const res = await NetworkCall.fetch(
-    UserRequest.VerifyOTP(values, auth.loggedInUser.token)
+    UserRequest.VerifyOTP(values, auth?.token)
   )
     .then((value) => {
       return value;
@@ -370,7 +366,7 @@ export const GetAlgorithmsList = async () => {
   let state = store.getState();
   const { auth } = state;
   const res = await NetworkCall.fetch(
-    UserRequest.GetAlgorithmsList(auth.loggedInUser?.token)
+    UserRequest.GetAlgorithmsList(auth?.token)
   )
     .then((value) => {
       return value;
@@ -384,9 +380,7 @@ export const GetAlgorithmsList = async () => {
 export const GetLanguageList = async () => {
   let state = store.getState();
   const { auth } = state;
-  const res = await NetworkCall.fetch(
-    UserRequest.GetLanguageList(auth.loggedInUser?.token)
-  )
+  const res = await NetworkCall.fetch(UserRequest.GetLanguageList(auth?.token))
     .then((value) => {
       return value;
     })
@@ -399,10 +393,11 @@ export const GetLanguageList = async () => {
 // forgot password
 export const forgotPasswordSendOTP = async (values: object) => {
   try {
-    const authToken = await createToken();
-
     const res = await NetworkCall.fetch(
-      UserRequest.forgotPasswordSendOTP(values, authToken?.access_token)
+      UserRequest.forgotPasswordSendOTP(
+        values,
+        (getCookies() as any)?.loginToken
+      )
     );
 
     return res;
@@ -413,10 +408,11 @@ export const forgotPasswordSendOTP = async (values: object) => {
 
 export const forgotPasswordVerifyOTP = async (values: object) => {
   try {
-    const authToken = await createToken();
-
     const res = await NetworkCall.fetch(
-      UserRequest.forgotPasswordVerifyOTP(values, authToken?.access_token)
+      UserRequest.forgotPasswordVerifyOTP(
+        values,
+        (getCookies() as any)?.loginToken
+      )
     );
 
     return res;
@@ -427,10 +423,11 @@ export const forgotPasswordVerifyOTP = async (values: object) => {
 
 export const forgotPasswordUpdate = async (values: object) => {
   try {
-    const authToken = await createToken();
-
     const res = await NetworkCall.fetch(
-      UserRequest.forgotPasswordUpdatePassword(values, authToken?.access_token)
+      UserRequest.forgotPasswordUpdatePassword(
+        values,
+        (getCookies() as any)?.loginToken
+      )
     );
 
     return res;
@@ -444,7 +441,7 @@ export const addNickName = async (values: object) => {
   const { auth } = state;
 
   const res = await NetworkCall.fetch(
-    UserRequest.addNickName(values, auth.loggedInUser.token)
+    UserRequest.addNickName(values, auth?.token)
   )
     .then((value) => {
       return value;
@@ -457,12 +454,16 @@ export const addNickName = async (values: object) => {
 };
 
 export const getNickNameList = async () => {
-  let state = store.getState();
-  const { auth } = state;
+  let state = await store.getState();
 
-  const res = await NetworkCall.fetch(
-    UserRequest.getNickNameList(auth.loggedInUser?.token)
-  )
+  const cc: any = getCookies();
+
+  const { auth } = state,
+    tc = cc?.loginToken;
+
+  let token = auth?.token || tc;
+
+  const res = await NetworkCall.fetch(UserRequest.getNickNameList(token))
     .then((value) => {
       return value;
     })
@@ -477,7 +478,7 @@ export const updateNickName = async (values: object, id: string) => {
   const { auth } = state;
 
   const res = await NetworkCall.fetch(
-    UserRequest.updateNickName(values, auth.loggedInUser.token, id)
+    UserRequest.updateNickName(values, auth?.token, id)
   )
     .then((value) => {
       return value;
@@ -490,10 +491,8 @@ export const updateNickName = async (values: object, id: string) => {
 
 export const resendOTPForRegistration = async (values: object) => {
   try {
-    const authToken = await createToken();
-
     const res = await NetworkCall.fetch(
-      UserRequest.resendOTPForRegistration(values, authToken?.access_token)
+      UserRequest.resendOTPForRegistration(values)
     );
 
     return res;
@@ -507,7 +506,7 @@ export const getDirectSupportedCampsList = async () => {
   const { auth } = state;
 
   const res = await NetworkCall.fetch(
-    UserRequest.getDirectSupportedCampsList(auth?.loggedInUser?.token)
+    UserRequest.getDirectSupportedCampsList(auth?.token)
   )
     .then((value) => {
       return value;
@@ -524,7 +523,7 @@ export const removeSupportedCampsEntireTopic = async (body) => {
   const { auth } = state;
 
   const res = await NetworkCall.fetch(
-    UserRequest.removeSupportedCampsEntireTopic(body, auth.loggedInUser.token)
+    UserRequest.removeSupportedCampsEntireTopic(body, auth?.token)
   )
     .then((value) => {
       return value;
@@ -541,7 +540,7 @@ export const removeOrUpdateDirectSupportCamps = async (body) => {
   const { auth } = state;
 
   const res = await NetworkCall.fetch(
-    UserRequest.removeOrUpdateDirectSupportCamps(body, auth.loggedInUser.token)
+    UserRequest.removeOrUpdateDirectSupportCamps(body, auth?.token)
   )
     .then((value) => {
       return value;
@@ -556,9 +555,7 @@ export const addSupport = async (body) => {
   let state = store.getState();
   const { auth } = state;
 
-  const res = await NetworkCall.fetch(
-    UserRequest.addSupport(body, auth.loggedInUser.token)
-  )
+  const res = await NetworkCall.fetch(UserRequest.addSupport(body, auth?.token))
     .then((value) => {
       return value;
     })
@@ -574,7 +571,7 @@ export const addDelegateSupportCamps = async (body) => {
   const { auth } = state;
 
   const res = await NetworkCall.fetch(
-    UserRequest.addDelegatedSupport(body, auth.loggedInUser.token)
+    UserRequest.addDelegatedSupport(body, auth?.token)
   )
     .then((value) => {
       return value;
@@ -590,7 +587,7 @@ export const getDelegatedSupportCampsList = async () => {
   const { auth } = state;
 
   const res = await NetworkCall.fetch(
-    UserRequest.getDelegatedSupportCampsList(auth.loggedInUser?.token)
+    UserRequest.getDelegatedSupportCampsList(auth?.token)
   )
     .then((value) => {
       return value;
@@ -606,7 +603,7 @@ export const userSocialAccountsList = async () => {
 
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.userSocialAccountsList(state.auth.token)
+      UserRequest.userSocialAccountsList(state?.auth?.token)
     );
 
     return res;
@@ -620,7 +617,7 @@ export const userSocialAccountDelete = async (id: string) => {
 
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.userSocialAccountDelete(state.auth.token, id)
+      UserRequest.userSocialAccountDelete(state?.auth?.token, id)
     );
 
     return res;
@@ -633,7 +630,7 @@ export const socialLoginLinkUser = async (values: object) => {
   const state = store.getState();
 
   try {
-    let token = state.auth.token;
+    let token = state?.auth?.token;
 
     const res = await NetworkCall.fetch(
       UserRequest.SocialLinkUser(values, token)
@@ -665,7 +662,7 @@ export const deactivateUser = async (body: object) => {
 
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.UserDeactivate(body, state.auth.token)
+      UserRequest.UserDeactivate(body, state?.auth?.token)
     );
 
     return res;
@@ -679,7 +676,7 @@ export const uploadFile = async (body) => {
   const { auth } = state;
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.UploadFile(body, auth.loggedInUser?.token)
+      UserRequest.UploadFile(body, auth?.token)
     );
     return res;
   } catch (err) {
@@ -700,7 +697,7 @@ export const getUploadFileAndFolder = async () => {
   const { auth } = state;
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.GetUploadFileAndFolder(auth.loggedInUser?.token)
+      UserRequest.GetUploadFileAndFolder(auth?.token)
     );
     return res;
   } catch (err) {
@@ -720,7 +717,7 @@ export const createFolderApi = async (body) => {
   const { auth } = state;
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.CreateFolder(body, auth.loggedInUser?.token)
+      UserRequest.CreateFolder(body, auth?.token)
     );
     return res;
   } catch (err) {
@@ -741,7 +738,7 @@ export const deleteFolderApi = async (id) => {
   const { auth } = state;
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.DeleteFolder(id, auth.loggedInUser?.token)
+      UserRequest.DeleteFolder(id, auth?.token)
     );
     return res;
   } catch (err) {
@@ -762,7 +759,7 @@ export const deleteUploadFileApi = async (id) => {
   const { auth } = state;
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.DeleteUploadFile(id, auth.loggedInUser?.token)
+      UserRequest.DeleteUploadFile(id, auth?.token)
     );
     return res;
   } catch (err) {
@@ -783,7 +780,7 @@ export const getFileInsideFolderApi = async (id) => {
   const { auth } = state;
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.GetFileInsideAFolder(id, auth.loggedInUser?.token)
+      UserRequest.GetFileInsideAFolder(id, auth?.token)
     );
     return res;
   } catch (err) {
@@ -846,8 +843,6 @@ export const verifyEmailOnSocial = async (body) => {
 
     let payload = {
       ...res.data.user,
-      token: res.data.auth?.access_token,
-      refresh_token: res.data?.auth?.refresh_token,
     };
     document.cookie =
       "loginToken=" +
@@ -925,7 +920,7 @@ export const globalSearchUploadFiles = async (reqbody) => {
   const { auth } = state;
   try {
     const res = await NetworkCall.fetch(
-      UserRequest.GlobalSearchUploadedFiles(reqbody, auth.loggedInUser?.token)
+      UserRequest.GlobalSearchUploadedFiles(reqbody, auth?.token)
     );
     return res;
   } catch (err) {
